requirements-worker.txt

# Core Python
requests>=2.31.0
mysql-connector-python>=8.0.33

# Playwright
playwright>=1.49.0

# PDF Extraction libs
pdfplumber>=0.10.3
pytesseract>=0.3.10
PyPDF2>=3.0.1
tqdm>=4.66.1

# Celery / Redis
celery>=5.3.0
redis>=5.0.0



# ML / Relevancy
torch
scikit-learn
joblib
transformers
sentence-transformers